Fix V2 block reuse commit ordering
